[ISSUE-1254] Move Label APIs to jmap-extensions-api
